move entity_type to common models folder
