+++ lib/src/analyzers/lint_analyzer/metrics/metrics_factory.dart
@@ -1,3 +1,4 @@
+import '../../models/entity_type.dart';
 import 'metrics_list/cyclomatic_complexity/cyclomatic_complexity_metric.dart';
 import 'metrics_list/lines_of_code_metric.dart';
 import 'metrics_list/maximum_nesting_level/maximum_nesting_level_metric.dart';
@@ -5,7 +6,6 @@ import 'metrics_list/number_of_methods_metric.dart';
 import 'metrics_list/number_of_parameters_metric.dart';
 import 'metrics_list/source_lines_of_code/source_lines_of_code_metric.dart';
 import 'metrics_list/weight_of_class_metric.dart';
-import 'models/entity_type.dart';
 import 'models/metric.dart';
 
 final _implementedMetrics = <String, Metric Function(Map<String, Object>)>{
