+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/maximum_nesting_level/maximum_nesting_level_metric.dart
@@ -3,11 +3,11 @@ import 'package:analyzer/dart/ast/ast.dart';
 import '../../../../../utils/node_utils.dart';
 import '../../../../../utils/string_extension.dart';
 import '../../../../models/context_message.dart';
+import '../../../../models/entity_type.dart';
 import '../../../../models/internal_resolved_unit_result.dart';
 import '../../../../models/scoped_class_declaration.dart';
 import '../../../../models/scoped_function_declaration.dart';
 import '../../metric_utils.dart';
-import '../../models/entity_type.dart';
 import '../../models/function_metric.dart';
 import '../../models/metric_computation_result.dart';
 import '../../models/metric_documentation.dart';
