+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/number_of_parameters_metric.dart
@@ -1,10 +1,10 @@
 import 'package:analyzer/dart/ast/ast.dart';
 
+import '../../../models/entity_type.dart';
 import '../../../models/internal_resolved_unit_result.dart';
 import '../../../models/scoped_class_declaration.dart';
 import '../../../models/scoped_function_declaration.dart';
 import '../metric_utils.dart';
-import '../models/entity_type.dart';
 import '../models/function_metric.dart';
 import '../models/metric_computation_result.dart';
 import '../models/metric_documentation.dart';
