+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/weight_of_class_metric.dart
@@ -2,6 +2,7 @@ import 'package:analyzer/dart/ast/ast.dart';
 
 import '../../../../utils/node_utils.dart';
 import '../../../models/context_message.dart';
+import '../../../models/entity_type.dart';
 import '../../../models/function_type.dart';
 import '../../../models/internal_resolved_unit_result.dart';
 import '../../../models/scoped_class_declaration.dart';
@@ -9,7 +10,6 @@ import '../../../models/scoped_function_declaration.dart';
 import '../../../utils/scope_utils.dart';
 import '../metric_utils.dart';
 import '../models/class_metric.dart';
-import '../models/entity_type.dart';
 import '../models/metric_computation_result.dart';
 import '../models/metric_documentation.dart';
