+++ lib/src/analyzers/lint_analyzer/metrics_analyzer.dart
@@ -13,6 +13,7 @@ import 'package:path/path.dart' as p;
 import '../../analyzer_plugin/plugin_utils.dart';
 import '../../config_builder/models/config.dart';
 import '../../utils/node_utils.dart';
+import '../models/entity_type.dart';
 import '../models/internal_resolved_unit_result.dart';
 import '../models/issue.dart';
 import '../models/report.dart';
@@ -26,7 +27,6 @@ import 'metrics/metric_utils.dart';
 import 'metrics/metrics_factory.dart';
 import 'metrics/metrics_list/cyclomatic_complexity/cyclomatic_complexity_metric.dart';
 import 'metrics/metrics_list/source_lines_of_code/source_code_visitor.dart';
-import 'metrics/models/entity_type.dart';
 import 'metrics/models/metric.dart';
 import 'metrics/models/metric_documentation.dart';
 import 'metrics/models/metric_value.dart';
