+++ lib/src/analyzers/models/entity_type.dart
